requirements.txt
reportlab
requests
python-docx
docx2pdf
python-dotenv
faker
pydantic
pandas
pytz
flask
python-dateutil